+++ frontend/src/components/nimmt/grid_row.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const GridRow = () => {
+  
+}
