+++ frontend/src/stylesheets/app.scss
@@ -9,8 +9,10 @@
 @import './profile/profile.scss';
 @import './lobby/lobby.scss';
 @import './profile/avatar.scss';
-@import './test/motion_framer.scss';
-@import './test/hand.scss';
+@import './game/motion_framer.scss';
+@import './game/cards.scss';
+@import './room/room.scss';
+@import './game/grid.scss';w
 
 /* http://meyerweb.com/eric/tools/css/reset/ 
    v2.0 | 20110126
