+++ frontend/src/stylesheets/fonts.scss
@@ -25,3 +25,10 @@
     font-weight: normal;
     font-style: normal;
 }
+
+@font-face {
+    font-family: "Marker";
+    src: url("../assets/fonts/marker.ttf");
+    font-weight: normal;
+    font-style: normal;
+}
