+++ frontend/src/stylesheets/game/grid.scss
@@ -1,0 +1,23 @@
+@import '../colors.scss';
+@import '../fonts.scss';
+
+.grid-container {
+  display: flex;
+}
+
+.game-grid {
+  display: grid;
+  margin-left: 12px;
+  margin-top: 12px;
+  width: fit-content;
+  grid-template-columns: 1fr;
+  // grid-template-rows: repeat(4, 1fr);
+  gap: 8px;
+}
+
+.blank-square {
+  height: 140px;
+  width: 100px;;
+  background-color: $champ;
+}
+
