+++ frontend/src/stylesheets/game/motion_framer.scss
