+++ frontend/src/stylesheets/room/room.scss
@@ -1,0 +1,3 @@
+@import '../colors.scss';
+@import '../fonts.scss';
+
